requirements.txt
langchain~=0.1.12
openai
streamlit~=1.32.2
python-dotenv~=1.0.1